Remove PHP 8.1 tests
These are unnecessary and expected to fail. While they're nice to have
for visibility, in GitHub actions the failure (with `continue-on-error`)
flags the overall commit or PR as a failure, which a red 'x'. This is
inconvenient, and having an accurate status check representation with a
green check is far more useful than testing a PHP version we don't
currently care about.

Many people experience the same issue, e.g.
<https://github.com/actions/toolkit/issues/399>